requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.3
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-decouple>=3.8
Pillow>=10.0
